Improve adaptiveness of custom bottom navbar
